refactor(assay): extract inventory.rs module
